+++ src/main/kotlin/ar/edu/unq/ttip/sportbook/persistence/entity/user/AuthToken.kt
@@ -1,0 +1,3 @@
+package ar.edu.unq.ttip.sportbook.persistence.entity.user
+
+data class AuthToken(val token: String, val expiresAt: Long)
