+++ src/main/kotlin/ar/edu/unq/ttip/sportbook/persistence/entity/user/Sport.kt
@@ -1,5 +1,5 @@
 package ar.edu.unq.ttip.sportbook.persistence.entity.user
 
 enum class Sport {
-    FOOTBALL, VOLLEY, PADDLE, UNKNOWN
+    FOOTBALL, VOLLEY, PADDLE
 }
